a start at doozer-based drain configuration
modifying doozer triggers logyard to start/stop drains. however since
zmq's "send" blocks by default, we are forced to use
http://api.zeromq.org/master:zmq-poll with channels; this requires some
effort, so now commiting whatever we got so far with debugging lines to
a branch.

this commit also includes the NewClientGlobal api that shares zeromq
context (as a precursor to using zmq_poll) in the same process.

http://bugs.activestate.com/show_bug.cgi?id=95916