Bug #98354 - warn only for the first retry
if a drain runs successfully for 20mins, then retry stats will be
resetted, leading to next disconnection and first retry logged as
WARN. all other retries are logged as INFO.

lastly, app drains will behave the same system drains in terms of
logging levels. previous to this commit, we never logged app drain
retry as a WARN. but now we do ... again, only for the first retry.