99381: disable startsecs for logyard
from http://supervisord.org/configuration.html

"The total number of seconds which the program needs to stay running
after a startup to consider the start successful. If the program does
not stay up for this many seconds after it has started, even if it
exits with an “expected” exit code (see exitcodes), the startup will
be considered a failure. Set to 0 to indicate that the program needn’t
stay running for any particular amount of time."

the default value of startsecs (1 second) is meaningless for the rest
of the processes (vcap processes) anyway, so we might as well disable
it for logyard. app_mdns already has startsecs disabled.