apptail: get app-specific logs from env var instead of stackato.yml
apptail now solely relies in $STACKATO_LOG_FILES as defined in /app/etc/droplet.env.json. this env var is set to the default list in the app image (changeable by the admin when creating a new app image), which can be added to, or overriden, by the application deployer. example:

+env:
+  STACKATO_LOG_FILES: "cust=/app/app/custom.log:$STACKATO_LOG_FILES"